src/test/java/ru/druzhininyy/leetcode/exercises/algorithms/problem0443/SolutionTest.java
package ru.druzhininyy.leetcode.exercises.algorithms.problem0443;

import com.fasterxml.jackson.core.type.TypeReference;
import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.io.IOException;
import java.util.*;
import java.util.stream.Stream;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.*;

@DisplayName("Test: 443. String Compression.")
public class SolutionTest {

    @Getter
    @Setter
    @ToString
    public static class TestCase {
        public char[] chars;
        public int expectedReturnValue;
        public char[] expectedCompressedString;
    }

    private static final String SOURCE_FILE = "problem0443.json";
    public static List<TestCase> testCases;

    @BeforeAll
    public static void loadTestCases() {
        ObjectMapper objectMapper = new ObjectMapper();
        try {
            testCases = objectMapper.readValue(
                    SolutionTest.class.getClassLoader().getResourceAsStream(SOURCE_FILE),
                    new TypeReference<>() {
                    });
        } catch (IOException e) {
            System.out.println("Error while loading data from " + SOURCE_FILE);
            System.out.println(e.getMessage());
        }
    }

    private static Stream<Arguments> provideTestsArguments() {
        return testCases.stream().map(Arguments::of);
    }

    @ParameterizedTest
    @MethodSource("provideTestsArguments")
    public void runTestCases(TestCase testCase) {
        System.out.println(testCase);

        var actual = Solution.compress(testCase.getChars());

        assertThat(Arrays.copyOfRange(testCase.getChars(), 0, actual), is(testCase.getExpectedCompressedString()));
        assertThat(actual, is(testCase.getExpectedReturnValue()));
    }

}










